Allow kebab-cased component files
